Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
openai==1.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
yt-dlp==2023.11.16
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1
alembic==1.12.1
pytest==7.4.3
httpx==0.24.1
python-slugify==8.0.1
PyJWT==2.8.0

# Test dependencies
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0